Stretch modifications
-FriendForm is now stateful and handles change and submit
-implemented routing
