+++ friends/src/App.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { Route } from 'react-router-dom';
 import axios from 'axios';
 import './App.css';
 
@@ -11,9 +12,6 @@ class App extends Component {
     this.state = {
       friends: [],
       isLoading: true,
-      name: '',
-      age: '',
-      email: '',
     }
   }
 
@@ -27,37 +25,11 @@ class App extends Component {
           });
   }
 
-  handleInputChange = event => {
-    this.setState({[event.target.name]: event.target.value});
-  }
-
-  handleSubmit = event => {
-    event.preventDefault();
-    const newFriend = {
-      name: this.state.name,
-      age: parseInt(this.state.age, 10),
-      email: this.state.email,
-    };
-    axios.post('http://localhost:5000/friends', newFriend)
-          .then( res =>
-            this.setState({name: '', age: '', email: '', friends: res.data})
-          );
-  }
-
   render() {
     return (
       <div className="App">
-        {this.state.isLoading ?
-          <h1>Loading Friends, one moment please...</h1> :
-          <FriendList friends={this.state.friends} />
-        }
-        <FriendForm
-          input={this.handleInputChange}
-          submit={this.handleSubmit}
-          name={this.state.name}
-          age={this.state.age}
-          email={this.state.email}
-        />
+        <Route exact path="/" render={props=> <FriendList {...props} friends={this.state.friends} loading={this.state.isLoading} />} />
+        <Route path="/new" component={FriendForm} />
       </div>
     );
   }
