+++ friends/src/components/FriendList.js
@@ -5,7 +5,10 @@ import FriendDisplay from './FriendDisplay';
 const FriendList = props => {
   return (
     <div>
-      {props.friends.map(friend => <FriendDisplay key={friend.id} {...friend} />)}
+      {props.isLoading ? 
+        <h1>Loading Friends, please hold...</h1> :
+        props.friends.map(friend => <FriendDisplay key={friend.id} {...friend} />)
+      }
     </div>
   );
 }
